requirements.txt
flask==3.0.3
pyqrcode==1.2.1
pypng==0.20220715.0
opencv-python==4.10.0.84
pillow==10.1.0
